+++ src/main/java/com/souldealers/crowdtracebackend/modules/identity/UserRoles.java
@@ -1,6 +1,7 @@
 package com.souldealers.crowdtracebackend.modules.identity;
 
 public enum UserRoles {
-    USER,
-    ADMIN
+    REGISTERED_USER,
+    MODERATOR,
+    SUPER_ADMIN
 }