+++ src/main/java/com/souldealers/crowdtracebackend/modules/identity/UserStatus.java
@@ -2,8 +2,7 @@ package com.souldealers.crowdtracebackend.modules.identity;
 
 public enum UserStatus {
     ACTIVE,
+    SUSPENDED,
     INACTIVE,
-    BLOCKED,
-    LOCKED,
-    VERIFIED
+    DEACTIVATED
 }