+++ src/main/java/com/souldealers/crowdtracebackend/modules/identity/internal/repository/UserRepository.java
@@ -1,12 +1,16 @@
 package com.souldealers.crowdtracebackend.modules.identity.internal.repository;
 
 import com.souldealers.crowdtracebackend.modules.identity.internal.model.User;
+import org.jspecify.annotations.NullMarked;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.UUID;
+import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, UUID> {
+public interface UserRepository extends JpaRepository<User, Long> {
+    @NullMarked
     Page<User> findAll(Pageable pageable);
+
+    Optional<User> findByEmail(String email);
 }